requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
openai
azure-cosmos
azure-search-documents
azure-storage-blob
azure-servicebus
pypdf
